+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/ListWorksetViewDto.java
@@ -1,0 +1,60 @@
+package uk.ac.bbsrc.tgac.miso.dto;
+
+public class ListWorksetViewDto {
+
+  private long id;
+  private String alias;
+  private int itemCount;
+  private String description;
+  private String creator;
+  private String lastModified;
+
+  public long getId() {
+    return id;
+  }
+
+  public void setId(long id) {
+    this.id = id;
+  }
+
+  public String getAlias() {
+    return alias;
+  }
+
+  public void setAlias(String alias) {
+    this.alias = alias;
+  }
+
+  public int getItemCount() {
+    return itemCount;
+  }
+
+  public void setItemCount(int itemCount) {
+    this.itemCount = itemCount;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+
+  public void setDescription(String description) {
+    this.description = description;
+  }
+
+  public String getCreator() {
+    return creator;
+  }
+
+  public void setCreator(String creator) {
+    this.creator = creator;
+  }
+
+  public String getLastModified() {
+    return lastModified;
+  }
+
+  public void setLastModified(String lastModified) {
+    this.lastModified = lastModified;
+  }
+
+}