+++ sqlstore/src/it/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateWorksetDaoIT.java
@@ -2,7 +2,7 @@ package uk.ac.bbsrc.tgac.miso.persistence.impl;
 
 import static org.junit.Assert.*;
 
-import java.io.IOException;
+import java.util.List;
 
 import org.hibernate.SessionFactory;
 import org.junit.Before;
@@ -13,7 +13,6 @@ import uk.ac.bbsrc.tgac.miso.AbstractDAOTest;
 import uk.ac.bbsrc.tgac.miso.core.data.Sample;
 import uk.ac.bbsrc.tgac.miso.core.data.Workset;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleImpl;
-import uk.ac.bbsrc.tgac.miso.core.util.PaginationFilter;
 
 public class HibernateWorksetDaoIT extends AbstractDAOTest {
 
@@ -86,26 +85,21 @@ public class HibernateWorksetDaoIT extends AbstractDAOTest {
   }
 
   @Test
-  public void testSearchByCreator() throws IOException {
-    testSearch(PaginationFilter.user("admin", true));
+  public void testListBySample() throws Exception {
+    List<Workset> results = sut.listBySample(1L);
+    assertEquals(2, results.size());
   }
 
   @Test
-  public void testSearchByModifier() throws IOException {
-    testSearch(PaginationFilter.user("admin", false));
+  public void testListByLibrary() throws Exception {
+    List<Workset> results = sut.listByLibrary(1L);
+    assertEquals(1, results.size());
   }
 
-  /**
-   * Verifies Hibernate mappings by ensuring that no exception is thrown by a search
-   * 
-   * @param filter the search filter
-   * @throws IOException
-   */
-  private void testSearch(PaginationFilter filter) throws IOException {
-    // verify Hibernate mappings by ensuring that no exception is thrown
-    assertNotNull(sut.list(err -> {
-      throw new RuntimeException(err);
-    }, 0, 10, true, "id", filter));
+  @Test
+  public void testListByLibraryAliquot() throws Exception {
+    List<Workset> results = sut.listByLibraryAliquot(1L);
+    assertEquals(1, results.size());
   }
 
 }