+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernateWorksetDao.java
@@ -1,26 +1,20 @@
 package uk.ac.bbsrc.tgac.miso.persistence.impl;
 
-import java.util.Collections;
 import java.util.List;
 
-import org.hibernate.Criteria;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.criterion.MatchMode;
 import org.hibernate.criterion.Restrictions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import uk.ac.bbsrc.tgac.miso.core.data.Workset;
-import uk.ac.bbsrc.tgac.miso.core.util.DateType;
 import uk.ac.bbsrc.tgac.miso.persistence.WorksetStore;
 
 @Repository
 @Transactional(rollbackFor = Exception.class)
-public class HibernateWorksetDao implements WorksetStore, HibernatePaginatedDataSource<Workset> {
-
-  private static final String[] SEARCH_PROPERTIES = new String[] { "alias", "description" };
+public class HibernateWorksetDao implements WorksetStore {
 
   @Autowired
   private SessionFactory sessionFactory;
@@ -29,7 +23,6 @@ public class HibernateWorksetDao implements WorksetStore, HibernatePaginatedData
     this.sessionFactory = sessionFactory;
   }
 
-  @Override
   public Session currentSession() {
     return sessionFactory.getCurrentSession();
   }
@@ -44,18 +37,6 @@ public class HibernateWorksetDao implements WorksetStore, HibernatePaginatedData
     return (Workset) currentSession().createCriteria(Workset.class)
         .add(Restrictions.eq("alias", alias))
         .uniqueResult();
-  }
-
-  @Override
-  public List<Workset> listBySearch(String query) {
-    if (query == null) {
-      throw new NullPointerException("No query string provided");
-    }
-    @SuppressWarnings("unchecked")
-    List<Workset> results = currentSession().createCriteria(Workset.class)
-        .add(Restrictions.ilike("alias", query, MatchMode.START))
-        .list();
-    return results;
   }
 
   @Override
@@ -96,58 +77,6 @@ public class HibernateWorksetDao implements WorksetStore, HibernatePaginatedData
       currentSession().update(workset);
       return workset.getId();
     }
-  }
-
-  @Override
-  public String getFriendlyName() {
-    return "Workset";
-  }
-
-  @Override
-  public String getProjectColumn() {
-    return null;
-  }
-
-  @Override
-  public Class<? extends Workset> getRealClass() {
-    return Workset.class;
-  }
-
-  @Override
-  public String[] getSearchProperties() {
-    return SEARCH_PROPERTIES;
-  }
-
-  @Override
-  public Iterable<AliasDescriptor> listAliases() {
-    return Collections.emptyList();
-  }
-
-  @Override
-  public String propertyForDate(Criteria criteria, DateType type) {
-    switch (type) {
-    case ENTERED:
-      return "created";
-    case UPDATE:
-      return "lastModified";
-    default:
-      return null;
-    }
-  }
-
-  @Override
-  public String propertyForId() {
-    return "id";
-  }
-
-  @Override
-  public String propertyForSortColumn(String original) {
-    return original;
-  }
-
-  @Override
-  public String propertyForUser(boolean creator) {
-    return creator ? "creator" : "lastModifier";
   }
 
 }