+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 
 Changes:
 
-
+  * Improved performace of Worksets list page
 
 # 1.1.0